requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
pytorch-lightning
matplotlib
scipy
scikit-learn
tensorboard
huggingface-hub[torch]>=0.22
jaxtyping
open3d
albumentations==0.4.3
opencv-python
pudb==2019.2
imageio==2.36.1
imageio-ffmpeg==0.4.2
omegaconf
test-tube>=0.7.5
streamlit>=0.73.1
einops
torch-fidelity==0.3.0
transformers==4.48.0
kornia==0.6
webdataset==0.2.5
torchmetrics
fire==0.4.0
gradio==3.21.0
diffusers
datasets[vision]==2.4.0
carvekit-colab==4.1.0
rich>=13.3.2
lovely-numpy>=0.2.8
lovely-tensors>=0.1.14
plotly==5.13.1
ninja
accelerate
ipdb
tqdm
wandb
open-clip-torch
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git
trimesh
icecream
pyglet<2
pyrender
roma
lpips

# for checkpoints download
s5cmd
gdown